% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/care.R
\name{care}
\alias{care}
\title{Estimate Precision Matrix for Compositional Data}
\usage{
care(x, lam)
}
\arguments{
\item{x}{n x p composition data matrix (row/column is sample/variable).}

\item{lam}{tuning parameter(s). It is either a vector or a single number. If it is a vector, each value in the vector corresponds to the parameter of a column; if it is a single number, this number is used as the parameter for every column.}
}
\value{
a list containing following components:

\item{prec} the estimator of precision matrix under the user-specified tuning parameters;
\item{lam} the given tuning parameter for each column.
}
\description{
\code{care} is used to estimate the precision matrix of the basis data based on the compositional data.
}
