% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sample_test.R
\name{cd.test}
\alias{cd.test}
\title{Two Samples Test for Compositional Data}
\usage{
cd.test(x1, x2, paired = FALSE)
}
\arguments{
\item{x1}{a numeric vector of compositional data values}

\item{x2}{another numeric vector of compositional data values}

\item{paired}{logical indicator for whether you want a paired test}
}
\value{
the p-value for the test
}
\description{
\code{cd.test} is used to assess whether there are significant differences between two compositional data samples.
}
