% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coat.R
\name{coat}
\alias{coat}
\title{Covariance Estimation for Compositional Data}
\usage{
coat(x, lam, nlam = 100, soft = 1)
}
\arguments{
\item{x}{n x p composition data matrix (row/column is sample/variable)}

\item{lam}{tuning parameter for thresholding}

\item{soft}{indicator for thresholding method: 1 for soft thresholding (default), 0 for hard thresholding}
}
\value{
a list containing following components:
\item{sigma}{covariance matrix }
\item{corr}{correlation matrx derived from the covariance matrix}
\item{lam}{the tuning parameter used}
}
\description{
\code{coat} is used to estimate the covariance matrix of basis data based on the compositional data.
}
