% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/care.R
\name{cv.care}
\alias{cv.care}
\title{Select Tunning Parameter for CARE Via Cross-Validation}
\usage{
cv.care(x, nlam = 50, minlam = 0.01, nfold = 5)
}
\arguments{
\item{x}{n x p composition data matrix (row/column is sample/variable)}

\item{nlam}{the number of tuning parameters(50 in default)}

\item{minlam}{the minimal tuning parameter(0.01 in default)}

\item{nfold}{the number of sample splitting(5 in default)}
}
\value{
a list containing following components:
\item{prec} the estimator of precision matrix under the selected tuning parameters;
\item{lam} the selected tuning parameter for each column.
}
\description{
\code{cv.care} uses k-fold cross-validation to find optimal tuning parameter for \code{care}.
}
