% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmm.R
\name{stab.cdmm}
\alias{stab.cdmm}
\title{Title}
\usage{
stab.cdmm(
  y,
  x,
  lam,
  nlam = 100,
  rlam = 1/nlam,
  nsample = 100,
  nsub = floor(0.5 * length(y)),
  constr = TRUE,
  seed = 213
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
