% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/care.R
\name{care}
\alias{care}
\title{Title}
\usage{
care(x, lambda_vec, max_it = 50)
}
\arguments{
\item{x}{n x p composition data matrix (row/column is sample/variable).}

\item{lambda_vec}{a p length vector of lambda for each column.}

\item{max_it}{The solution process will stop when the number of iterations exceeds \code{max_it}(default 50).}
}
\value{
a list containing following components:
\item{Omega_hat} (p * p) matrix,
the estimator of Omega under the given tuning parameter;
\item{lambda_op}, (p * 1) vector,
the given tuning parameter for each column.
}
\description{
Title
}
