% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sample_test.R
\name{cd.test}
\alias{cd.test}
\title{Title}
\usage{
cd.test(x1, x2, paired = FALSE)
}
\arguments{
\item{x1}{a numeric vector of compositional data values}

\item{x2}{another numeric vector of compositional data values}

\item{paired}{a logical indicating whether you want a paired test}
}
\value{
the p-value for the test
}
\description{
Title
}
