% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmm.R
\name{cdmm}
\alias{cdmm}
\title{Title}
\usage{
cdmm(
  y,
  x,
  lam,
  nlam = 100,
  rlam = 1/nlam,
  mu = 1,
  std = TRUE,
  maxv = 0.4 * length(y),
  maxit = c(20, 50),
  tol = c(1e-04, 1e-07)
)
}
\arguments{
\item{tol}{}
}
\description{
Title
}
