% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coat.R
\name{coat}
\alias{coat}
\title{Covariance Estimation for Compositional Data Via Composition-Adjusted Thresholding}
\usage{
coat(x, lambda, soft = 1)
}
\arguments{
\item{x}{n x p composition data matrix (row/column is sample/variable)}

\item{lambda}{tuning parameter for thresholding}

\item{soft}{indicator for thresholding method: 1 for soft thresholding (default), 0 for hard thresholding}
}
\value{
a list containing following components:
\item{sigma}{covariance matrix }
\item{corr}{correlation matrx derived from the covariance matrix}
\item{lambda}{the tuning parameter used}
}
\description{
Covariance Estimation for Compositional Data Via Composition-Adjusted Thresholding
}
