% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coat.R
\name{cv.coat}
\alias{cv.coat}
\title{Select Tunning Parameter for COAT Via Cross-Validation}
\usage{
cv.coat(x, lam, nlam = 100, nfold = 5, foldid = NULL, soft = 1)
}
\arguments{
\item{x}{n x p composition data matrix (row/column is sample/variable)}

\item{lam}{tuning parameters for thresholding. The function will choose the optimized one in the parameter sequence.. If missing, a sequence will be automatically generated based on data.}

\item{nfold}{number of folds for cross-validation (default: 5)}

\item{soft}{indicator for thresholding method: 1 for soft thresholding (default), 0 for hard thresholding}
}
\value{
a list containing following components:
\item{sigma}{optimal covariance matrix selected via cross-validation}
\item{corr}{correlation matrix derived from the optimal covariance matrix}
\item{time}{execution time in seconds}
\item{lambda}{optimal tuning parameter selected via cross-validation}
}
\description{
\code{cv.coat} is used to choose tuning parameter \code{lam} for \code{coat} by the way of k-fold cross validation.
}
